@@ -1,3 +1,4 @@
+import contextlib
 import copy
 import datetime
 import decimal
@@ -6,8 +7,8 @@ import inspect
 import re
 import uuid
 import warnings
-from collections import OrderedDict
 from collections.abc import Mapping
+from enum import Enum
 
 from django.conf import settings
 from django.core.exceptions import ObjectDoesNotExist
@@ -15,7 +16,7 @@ from django.core.exceptions import ValidationError as DjangoValidationError
 from django.core.validators import (
     EmailValidator, MaxLengthValidator, MaxValueValidator, MinLengthValidator,
     MinValueValidator, ProhibitNullCharactersValidator, RegexValidator,
-    URLValidator, ip_address_validators
+    URLValidator
 )
 from django.forms import FilePathField as DjangoFilePathField
 from django.forms import ImageField as DjangoImageField
@@ -27,17 +28,20 @@ from django.utils.duration import duration_string
 from django.utils.encoding import is_protected_type, smart_str
 from django.utils.formats import localize_input, sanitize_separators
 from django.utils.ipv6 import clean_ipv6_address
-from django.utils.timezone import utc
 from django.utils.translation import gettext_lazy as _
-from pytz.exceptions import InvalidTimeError
 
-from rest_framework import (
-    ISO_8601, RemovedInDRF313Warning, RemovedInDRF314Warning
-)
+try:
+    import pytz
+except ImportError:
+    pytz = None
+
+from rest_framework import ISO_8601
+from rest_framework.compat import ip_address_validators
 from rest_framework.exceptions import ErrorDetail, ValidationError
 from rest_framework.settings import api_settings
 from rest_framework.utils import html, humanize_datetime, json, representation
 from rest_framework.utils.formatting import lazy_format
+from rest_framework.utils.timezone import valid_datetime
 from rest_framework.validators import ProhibitSurrogateCharactersValidator
 
 
@@ -63,6 +67,9 @@ def is_simple_callable(obj):
     """
     True if the object is a callable that takes no arguments.
     """
+    if not callable(obj):
+        return False
+
     # Bail early since we cannot inspect built-in function signatures.
     if inspect.isbuiltin(obj):
         raise BuiltinSignatureError(
@@ -109,27 +116,6 @@ def get_attribute(instance, attrs):
     return instance
 
 
-def set_value(dictionary, keys, value):
-    """
-    Similar to Python's built in `dictionary[key] = value`,
-    but takes a list of nested keys instead of a single key.
-
-    set_value({'a': 1}, [], {'b': 2}) -> {'a': 1, 'b': 2}
-    set_value({'a': 1}, ['x'], 2) -> {'a': 1, 'x': 2}
-    set_value({'a': 1}, ['x', 'y'], 2) -> {'a': 1, 'x': {'y': 2}}
-    """
-    if not keys:
-        dictionary.update(value)
-        return
-
-    for key in keys[:-1]:
-        if key not in dictionary:
-            dictionary[key] = {}
-        dictionary = dictionary[key]
-
-    dictionary[keys[-1]] = value
-
-
 def to_choices_dict(choices):
     """
     Convert choices into key/value dicts.
@@ -142,7 +128,7 @@ def to_choices_dict(choices):
     # choices = [1, 2, 3]
     # choices = [(1, 'First'), (2, 'Second'), (3, 'Third')]
     # choices = [('Category', ((1, 'First'), (2, 'Second'))), (3, 'Third')]
-    ret = OrderedDict()
+    ret = {}
     for choice in choices:
         if not isinstance(choice, (list, tuple)):
             # single choice
@@ -165,7 +151,7 @@ def flatten_choices_dict(choices):
     flatten_choices_dict({1: '1st', 2: '2nd'}) -> {1: '1st', 2: '2nd'}
     flatten_choices_dict({'Group': {1: '1st', 2: '2nd'}}) -> {1: '1st', 2: '2nd'}
     """
-    ret = OrderedDict()
+    ret = {}
     for key, value in choices.items():
         if isinstance(value, dict):
             # grouped choices (category, sub choices)
@@ -263,16 +249,6 @@ class CreateOnlyDefault:
         if is_update:
             raise SkipField()
         if callable(self.default):
-            if hasattr(self.default, 'set_context'):
-                warnings.warn(
-                    "Method `set_context` on defaults is deprecated and will "
-                    "no longer be called starting with 3.13. Instead set "
-                    "`requires_context = True` on the class, and accept the "
-                    "context as an additional argument.",
-                    RemovedInDRF313Warning, stacklevel=2
-                )
-                self.default.set_context(self)
-
             if getattr(self.default, 'requires_context', False):
                 return self.default(serializer_field)
             else:
@@ -365,6 +341,10 @@ class Field:
             messages.update(getattr(cls, 'default_error_messages', {}))
         messages.update(error_messages or {})
         self.error_messages = messages
+
+    # Allow generic typing checking for fields.
+    def __class_getitem__(cls, *args, **kwargs):
+        return cls
 
     def bind(self, field_name, parent):
         """
@@ -502,16 +482,6 @@ class Field:
             # No default, or this is a partial update.
             raise SkipField()
         if callable(self.default):
-            if hasattr(self.default, 'set_context'):
-                warnings.warn(
-                    "Method `set_context` on defaults is deprecated and will "
-                    "no longer be called starting with 3.13. Instead set "
-                    "`requires_context = True` on the class, and accept the "
-                    "context as an additional argument.",
-                    RemovedInDRF313Warning, stacklevel=2
-                )
-                self.default.set_context(self)
-
             if getattr(self.default, 'requires_context', False):
                 return self.default(self)
             else:
@@ -576,16 +546,6 @@ class Field:
         """
         errors = []
         for validator in self.validators:
-            if hasattr(validator, 'set_context'):
-                warnings.warn(
-                    "Method `set_context` on validators is deprecated and will "
-                    "no longer be called starting with 3.13. Instead set "
-                    "`requires_context = True` on the class, and accept the "
-                    "context as an additional argument.",
-                    RemovedInDRF313Warning, stacklevel=2
-                )
-                validator.set_context(self)
-
             try:
                 if getattr(validator, 'requires_context', False):
                     validator(value, self)
@@ -703,60 +663,58 @@ class BooleanField(Field):
     default_empty_html = False
     initial = False
     TRUE_VALUES = {
-        't', 'T',
-        'y', 'Y', 'yes', 'Yes', 'YES',
-        'true', 'True', 'TRUE',
-        'on', 'On', 'ON',
-        '1', 1,
-        True
+        't',
+        'y',
+        'yes',
+        'true',
+        'on',
+        '1',
+        1,
+        True,
     }
     FALSE_VALUES = {
-        'f', 'F',
-        'n', 'N', 'no', 'No', 'NO',
-        'false', 'False', 'FALSE',
-        'off', 'Off', 'OFF',
-        '0', 0, 0.0,
-        False
+        'f',
+        'n',
+        'no',
+        'false',
+        'off',
+        '0',
+        0,
+        0.0,
+        False,
     }
-    NULL_VALUES = {'null', 'Null', 'NULL', '', None}
-
-    def to_internal_value(self, data):
-        try:
-            if data in self.TRUE_VALUES:
-                return True
-            elif data in self.FALSE_VALUES:
-                return False
-            elif data in self.NULL_VALUES and self.allow_null:
-                return None
-        except TypeError:  # Input is an unhashable type
-            pass
-        self.fail('invalid', input=data)
-
-    def to_representation(self, value):
-        if value in self.TRUE_VALUES:
-            return True
-        elif value in self.FALSE_VALUES:
-            return False
-        if value in self.NULL_VALUES and self.allow_null:
-            return None
-        return bool(value)
-
-
-class NullBooleanField(BooleanField):
-    initial = None
+    NULL_VALUES = {'null', '', None}
 
     def __init__(self, **kwargs):
-        warnings.warn(
-            "The `NullBooleanField` is deprecated and will be removed starting "
-            "with 3.14. Instead use the `BooleanField` field and set "
-            "`allow_null=True` which does the same thing.",
-            RemovedInDRF314Warning, stacklevel=2
-        )
-
-        assert 'allow_null' not in kwargs, '`allow_null` is not a valid option.'
-        kwargs['allow_null'] = True
-
+        if kwargs.get('allow_null', False):
+            self.default_empty_html = None
+            self.initial = None
         super().__init__(**kwargs)
+
+    @staticmethod
+    def _lower_if_str(value):
+        if isinstance(value, str):
+            return value.lower()
+        return value
+
+    def to_internal_value(self, data):
+        with contextlib.suppress(TypeError):
+            if self._lower_if_str(data) in self.TRUE_VALUES:
+                return True
+            elif self._lower_if_str(data) in self.FALSE_VALUES:
+                return False
+            elif self._lower_if_str(data) in self.NULL_VALUES and self.allow_null:
+                return None
+        self.fail("invalid", input=data)
+
+    def to_representation(self, value):
+        if self._lower_if_str(value) in self.TRUE_VALUES:
+            return True
+        elif self._lower_if_str(value) in self.FALSE_VALUES:
+            return False
+        if self._lower_if_str(value) in self.NULL_VALUES and self.allow_null:
+            return None
+        return bool(value)
 
 
 # String types...
@@ -907,7 +865,7 @@ class IPAddressField(CharField):
         self.protocol = protocol.lower()
         self.unpack_ipv4 = (self.protocol == 'both')
         super().__init__(**kwargs)
-        validators, error_message = ip_address_validators(protocol, self.unpack_ipv4)
+        validators = ip_address_validators(protocol, self.unpack_ipv4)
         self.validators.extend(validators)
 
     def to_internal_value(self, data):
@@ -968,7 +926,8 @@ class FloatField(Field):
         'invalid': _('A valid number is required.'),
         'max_value': _('Ensure this value is less than or equal to {max_value}.'),
         'min_value': _('Ensure this value is greater than or equal to {min_value}.'),
-        'max_string_length': _('String value too large.')
+        'max_string_length': _('String value too large.'),
+        'overflow': _('Integer value too large to convert to float')
     }
     MAX_STRING_LENGTH = 1000  # Guard against malicious string inputs.
 
@@ -994,6 +953,8 @@ class FloatField(Field):
             return float(data)
         except (TypeError, ValueError):
             self.fail('invalid')
+        except OverflowError:
+            self.fail('overflow')
 
     def to_representation(self, value):
         return float(value)
@@ -1012,10 +973,11 @@ class DecimalField(Field):
     MAX_STRING_LENGTH = 1000  # Guard against malicious string inputs.
 
     def __init__(self, max_digits, decimal_places, coerce_to_string=None, max_value=None, min_value=None,
-                 localize=False, rounding=None, **kwargs):
+                 localize=False, rounding=None, normalize_output=False, **kwargs):
         self.max_digits = max_digits
         self.decimal_places = decimal_places
         self.localize = localize
+        self.normalize_output = normalize_output
         if coerce_to_string is not None:
             self.coerce_to_string = coerce_to_string
         if self.localize:
@@ -1023,6 +985,11 @@ class DecimalField(Field):
 
         self.max_value = max_value
         self.min_value = min_value
+
+        if self.max_value is not None and not isinstance(self.max_value, (int, decimal.Decimal)):
+            warnings.warn("max_value should be an integer or Decimal instance.")
+        if self.min_value is not None and not isinstance(self.min_value, (int, decimal.Decimal)):
+            warnings.warn("min_value should be an integer or Decimal instance.")
 
         if self.max_digits is not None and self.decimal_places is not None:
             self.max_whole_digits = self.max_digits - self.decimal_places
@@ -1128,6 +1095,9 @@ class DecimalField(Field):
 
         quantized = self.quantize(value)
 
+        if self.normalize_output:
+            quantized = quantized.normalize()
+
         if not coerce_to_string:
             return quantized
         if self.localize:
@@ -1177,7 +1147,7 @@ class DateTimeField(Field):
         When `self.default_timezone` is `None`, always return naive datetimes.
         When `self.default_timezone` is not `None`, always return aware datetimes.
         """
-        field_timezone = getattr(self, 'timezone', self.default_timezone())
+        field_timezone = self.timezone if hasattr(self, 'timezone') else self.default_timezone()
 
         if field_timezone is not None:
             if timezone.is_aware(value):
@@ -1186,11 +1156,18 @@ class DateTimeField(Field):
                 except OverflowError:
                     self.fail('overflow')
             try:
-                return timezone.make_aware(value, field_timezone)
-            except InvalidTimeError:
-                self.fail('make_aware', timezone=field_timezone)
+                dt = timezone.make_aware(value, field_timezone)
+                # When the resulting datetime is a ZoneInfo instance, it won't necessarily
+                # throw given an invalid datetime, so we need to specifically check.
+                if not valid_datetime(dt):
+                    self.fail('make_aware', timezone=field_timezone)
+                return dt
+            except Exception as e:
+                if pytz and isinstance(e, pytz.exceptions.InvalidTimeError):
+                    self.fail('make_aware', timezone=field_timezone)
+                raise e
         elif (field_timezone is None) and timezone.is_aware(value):
-            return timezone.make_naive(value, utc)
+            return timezone.make_naive(value, datetime.timezone.utc)
         return value
 
     def default_timezone(self):
@@ -1206,19 +1183,14 @@ class DateTimeField(Field):
             return self.enforce_timezone(value)
 
         for input_format in input_formats:
-            if input_format.lower() == ISO_8601:
-                try:
+            with contextlib.suppress(ValueError, TypeError):
+                if input_format.lower() == ISO_8601:
                     parsed = parse_datetime(value)
                     if parsed is not None:
                         return self.enforce_timezone(parsed)
-                except (ValueError, TypeError):
-                    pass
-            else:
-                try:
-                    parsed = self.datetime_parser(value, input_format)
-                    return self.enforce_timezone(parsed)
-                except (ValueError, TypeError):
-                    pass
+
+                parsed = self.datetime_parser(value, input_format)
+                return self.enforce_timezone(parsed)
 
         humanized_format = humanize_datetime.datetime_formats(input_formats)
         self.fail('invalid', format=humanized_format)
@@ -1376,6 +1348,7 @@ class DurationField(Field):
         'invalid': _('Duration has wrong format. Use one of these formats instead: {format}.'),
         'max_value': _('Ensure this value is less than or equal to {max_value}.'),
         'min_value': _('Ensure this value is greater than or equal to {min_value}.'),
+        'overflow': _('The number of days must be between {min_days} and {max_days}.'),
     }
 
     def __init__(self, **kwargs):
@@ -1394,7 +1367,10 @@ class DurationField(Field):
     def to_internal_value(self, value):
         if isinstance(value, datetime.timedelta):
             return value
-        parsed = parse_duration(str(value))
+        try:
+            parsed = parse_duration(str(value))
+        except OverflowError:
+            self.fail('overflow', min_days=datetime.timedelta.min.days, max_days=datetime.timedelta.max.days)
         if parsed is not None:
             return parsed
         self.fail('invalid', format='[DD] [HH:[MM:]]ss[.uuuuuu]')
@@ -1424,7 +1400,8 @@ class ChoiceField(Field):
     def to_internal_value(self, data):
         if data == '' and self.allow_blank:
             return ''
-
+        if isinstance(data, Enum) and str(data) != str(data.value):
+            data = data.value
         try:
             return self.choice_strings_to_values[str(data)]
         except KeyError:
@@ -1433,6 +1410,8 @@ class ChoiceField(Field):
     def to_representation(self, value):
         if value in ('', None):
             return value
+        if isinstance(value, Enum) and str(value) != str(value.value):
+            value = value.value
         return self.choice_strings_to_values.get(str(value), value)
 
     def iter_options(self):
@@ -1456,7 +1435,7 @@ class ChoiceField(Field):
         # Allows us to deal with eg. integer choices while supporting either
         # integer or string input, but still get the correct datatype out.
         self.choice_strings_to_values = {
-            str(key): key for key in self.choices
+            str(key.value) if isinstance(key, Enum) and str(key) != str(key.value) else str(key): key for key in self.choices
         }
 
     choices = property(_get_choices, _set_choices)
@@ -1491,6 +1470,8 @@ class MultipleChoiceField(ChoiceField):
             self.fail('empty')
 
         return {
+            # Arguments for super() are needed because of scoping inside
+            # comprehensions.
             super(MultipleChoiceField, self).to_internal_value(item)
             for item in data
         }
@@ -1515,6 +1496,7 @@ class FilePathField(ChoiceField):
             allow_folders=allow_folders, required=required
         )
         kwargs['choices'] = field.choices
+        kwargs['required'] = required
         super().__init__(**kwargs)
 
 
@@ -1673,13 +1655,15 @@ class ListField(Field):
 
     def run_child_validation(self, data):
         result = []
-        errors = OrderedDict()
+        errors = {}
 
         for idx, item in enumerate(data):
             try:
                 result.append(self.child.run_validation(item))
             except ValidationError as e:
                 errors[idx] = e.detail
+            except DjangoValidationError as e:
+                errors[idx] = get_error_detail(e)
 
         if not errors:
             return result
@@ -1735,7 +1719,7 @@ class DictField(Field):
 
     def run_child_validation(self, data):
         result = {}
-        errors = OrderedDict()
+        errors = {}
 
         for key, value in data.items():
             key = str(key)
@@ -1837,6 +1821,7 @@ class HiddenField(Field):
     constraint on a pair of fields, as we need some way to include the date in
     the validated data.
     """
+
     def __init__(self, **kwargs):
         assert 'default' in kwargs, 'default is a required argument.'
         kwargs['write_only'] = True
@@ -1860,12 +1845,13 @@ class SerializerMethodField(Field):
 
     For example:
 
-    class ExampleSerializer(self):
+    class ExampleSerializer(Serializer):
         extra_info = SerializerMethodField()
 
         def get_extra_info(self, obj):
             return ...  # Calculate some data to return.
     """
+
     def __init__(self, method_name=None, **kwargs):
         self.method_name = method_name
         kwargs['source'] = '*'
