@@ -4,8 +4,9 @@ Parsers are used to parse the content of incoming HTTP requests.
 They give us a generic way of being able to handle various media types
 on the request, such as form content or json encoded data.
 """
+
 import codecs
-from urllib import parse
+import contextlib
 
 from django.conf import settings
 from django.core.files.uploadhandler import StopFutureHandlers
@@ -13,8 +14,8 @@ from django.http import QueryDict
 from django.http.multipartparser import ChunkIter
 from django.http.multipartparser import \
     MultiPartParser as DjangoMultiPartParser
-from django.http.multipartparser import MultiPartParserError, parse_header
-from django.utils.encoding import force_str
+from django.http.multipartparser import MultiPartParserError
+from django.utils.http import parse_header_parameters
 
 from rest_framework import renderers
 from rest_framework.exceptions import ParseError
@@ -194,30 +195,12 @@ class FileUploadParser(BaseParser):
         Detects the uploaded file name. First searches a 'filename' url kwarg.
         Then tries to parse Content-Disposition header.
         """
-        try:
+        with contextlib.suppress(KeyError):
             return parser_context['kwargs']['filename']
-        except KeyError:
-            pass
 
-        try:
+        with contextlib.suppress(AttributeError, KeyError, ValueError):
             meta = parser_context['request'].META
-            disposition = parse_header(meta['HTTP_CONTENT_DISPOSITION'].encode())
-            filename_parm = disposition[1]
-            if 'filename*' in filename_parm:
-                return self.get_encoded_filename(filename_parm)
-            return force_str(filename_parm['filename'])
-        except (AttributeError, KeyError, ValueError):
-            pass
-
-    def get_encoded_filename(self, filename_parm):
-        """
-        Handle encoded filenames per RFC6266. See also:
-        https://tools.ietf.org/html/rfc2231#section-4
-        """
-        encoded_filename = force_str(filename_parm['filename*'])
-        try:
-            charset, lang, filename = encoded_filename.split('\'', 2)
-            filename = parse.unquote(filename)
-        except (ValueError, LookupError):
-            filename = force_str(filename_parm['filename'])
-        return filename
+            disposition, params = parse_header_parameters(meta['HTTP_CONTENT_DISPOSITION'])
+            if 'filename*' in params:
+                return params['filename*']
+            return params['filename']
