@@ -46,6 +46,17 @@ class OperandHolder(OperationHolderMixin):
         op2 = self.op2_class(*args, **kwargs)
         return self.operator_class(op1, op2)
 
+    def __eq__(self, other):
+        return (
+            isinstance(other, OperandHolder) and
+            self.operator_class == other.operator_class and
+            self.op1_class == other.op1_class and
+            self.op2_class == other.op2_class
+        )
+
+    def __hash__(self):
+        return hash((self.operator_class, self.op1_class, self.op2_class))
+
 
 class AND:
     def __init__(self, op1, op2):
@@ -78,8 +89,11 @@ class OR:
 
     def has_object_permission(self, request, view, obj):
         return (
-            self.op1.has_object_permission(request, view, obj) or
-            self.op2.has_object_permission(request, view, obj)
+            self.op1.has_permission(request, view)
+            and self.op1.has_object_permission(request, view, obj)
+        ) or (
+            self.op2.has_permission(request, view)
+            and self.op2.has_object_permission(request, view, obj)
         )
 
 
@@ -217,14 +231,14 @@ class DjangoModelPermissions(BasePermission):
         return view.queryset
 
     def has_permission(self, request, view):
+        if not request.user or (
+           not request.user.is_authenticated and self.authenticated_users_only):
+            return False
+
         # Workaround to ensure DjangoModelPermissions are not applied
         # to the root view when using DefaultRouter.
         if getattr(view, '_ignore_model_permissions', False):
             return True
-
-        if not request.user or (
-           not request.user.is_authenticated and self.authenticated_users_only):
-            return False
 
         queryset = self._queryset(view)
         perms = self.get_required_permissions(request.method, queryset.model)
