@@ -14,11 +14,11 @@ from contextlib import contextmanager
 
 from django.conf import settings
 from django.http import HttpRequest, QueryDict
-from django.http.multipartparser import parse_header
 from django.http.request import RawPostDataException
 from django.utils.datastructures import MultiValueDict
+from django.utils.http import parse_header_parameters
 
-from rest_framework import HTTP_HEADER_ENCODING, exceptions
+from rest_framework import exceptions
 from rest_framework.settings import api_settings
 
 
@@ -26,7 +26,7 @@ def is_form_media_type(media_type):
     """
     Return True if the media type is a valid form media type.
     """
-    base_media_type, params = parse_header(media_type.encode(HTTP_HEADER_ENCODING))
+    base_media_type, params = parse_header_parameters(media_type)
     return (base_media_type == 'application/x-www-form-urlencoded' or
             base_media_type == 'multipart/form-data')
 
@@ -186,6 +186,10 @@ class Request:
             self.method,
             self.get_full_path())
 
+    # Allow generic typing checking for requests.
+    def __class_getitem__(cls, *args, **kwargs):
+        return cls
+
     def _default_negotiator(self):
         return api_settings.DEFAULT_CONTENT_NEGOTIATION_CLASS()
 
@@ -213,7 +217,8 @@ class Request:
     @property
     def data(self):
         if not _hasattr(self, '_full_data'):
-            self._load_data_and_files()
+            with wrap_attributeerrors():
+                self._load_data_and_files()
         return self._full_data
 
     @property
@@ -413,22 +418,17 @@ class Request:
         to proxy it to the underlying HttpRequest object.
         """
         try:
-            return getattr(self._request, attr)
+            _request = self.__getattribute__("_request")
+            return getattr(_request, attr)
         except AttributeError:
-            return self.__getattribute__(attr)
-
-    @property
-    def DATA(self):
-        raise NotImplementedError(
-            '`request.DATA` has been deprecated in favor of `request.data` '
-            'since version 3.0, and has been fully removed as of version 3.2.'
-        )
+            raise AttributeError(f"'{self.__class__.__name__}' object has no attribute '{attr}'")
 
     @property
     def POST(self):
         # Ensure that request.POST uses our request parsing.
         if not _hasattr(self, '_data'):
-            self._load_data_and_files()
+            with wrap_attributeerrors():
+                self._load_data_and_files()
         if is_form_media_type(self.content_type):
             return self._data
         return QueryDict('', encoding=self._request._encoding)
@@ -439,15 +439,9 @@ class Request:
         # Different from the other two cases, which are not valid property
         # names on the WSGIRequest class.
         if not _hasattr(self, '_files'):
-            self._load_data_and_files()
+            with wrap_attributeerrors():
+                self._load_data_and_files()
         return self._files
-
-    @property
-    def QUERY_PARAMS(self):
-        raise NotImplementedError(
-            '`request.QUERY_PARAMS` has been deprecated in favor of `request.query_params` '
-            'since version 3.0, and has been fully removed as of version 3.2.'
-        )
 
     def force_plaintext_errors(self, value):
         # Hack to allow our exception handler to force choice of
