@@ -19,7 +19,9 @@ REST framework settings, checking for user settings first, then falling
 back to the defaults.
 """
 from django.conf import settings
-from django.test.signals import setting_changed
+# Import from `django.core.signals` instead of the official location
+# `django.test.signals` to avoid importing the test module unnecessarily.
+from django.core.signals import setting_changed
 from django.utils.module_loading import import_string
 
 from rest_framework import ISO_8601
@@ -114,7 +116,7 @@ DEFAULTS = {
     'COERCE_DECIMAL_TO_STRING': True,
     'UPLOADED_FILES_USE_URL': True,
 
-    # Browseable API
+    # Browsable API
     'HTML_SELECT_CUTOFF': 1000,
     'HTML_SELECT_CUTOFF_TEXT': "More than {count} items...",
 
