@@ -3,7 +3,6 @@
 import io
 from importlib import import_module
 
-import django
 from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured
 from django.core.handlers.wsgi import WSGIHandler
@@ -151,15 +150,19 @@ class APIRequestFactory(DjangoRequestFactory):
         """
         Encode the data returning a two tuple of (bytes, content_type)
         """
-
         if data is None:
-            return ('', content_type)
+            return (b'', content_type)
 
         assert format is None or content_type is None, (
             'You may not set both `format` and `content_type`.'
         )
 
         if content_type:
+            try:
+                data = self._encode_json(data, content_type)
+            except AttributeError:
+                pass
+
             # Content type specified explicitly, treat data as a raw bytestring
             ret = force_bytes(data, settings.DEFAULT_CHARSET)
 
@@ -277,7 +280,7 @@ class APIClient(APIRequestFactory, DjangoClient):
         """
         self.handler._force_user = user
         self.handler._force_token = token
-        if user is None:
+        if user is None and token is None:
             self.logout()  # Also clear any possible session info if required
 
     def request(self, **kwargs):
@@ -288,7 +291,7 @@ class APIClient(APIRequestFactory, DjangoClient):
     def get(self, path, data=None, follow=False, **extra):
         response = super().get(path, data=data, **extra)
         if follow:
-            response = self._handle_redirects(response, **extra)
+            response = self._handle_redirects(response, data=data, **extra)
         return response
 
     def post(self, path, data=None, format=None, content_type=None,
@@ -296,7 +299,7 @@ class APIClient(APIRequestFactory, DjangoClient):
         response = super().post(
             path, data=data, format=format, content_type=content_type, **extra)
         if follow:
-            response = self._handle_redirects(response, **extra)
+            response = self._handle_redirects(response, data=data, format=format, content_type=content_type, **extra)
         return response
 
     def put(self, path, data=None, format=None, content_type=None,
@@ -304,7 +307,7 @@ class APIClient(APIRequestFactory, DjangoClient):
         response = super().put(
             path, data=data, format=format, content_type=content_type, **extra)
         if follow:
-            response = self._handle_redirects(response, **extra)
+            response = self._handle_redirects(response, data=data, format=format, content_type=content_type, **extra)
         return response
 
     def patch(self, path, data=None, format=None, content_type=None,
@@ -312,7 +315,7 @@ class APIClient(APIRequestFactory, DjangoClient):
         response = super().patch(
             path, data=data, format=format, content_type=content_type, **extra)
         if follow:
-            response = self._handle_redirects(response, **extra)
+            response = self._handle_redirects(response, data=data, format=format, content_type=content_type, **extra)
         return response
 
     def delete(self, path, data=None, format=None, content_type=None,
@@ -320,7 +323,7 @@ class APIClient(APIRequestFactory, DjangoClient):
         response = super().delete(
             path, data=data, format=format, content_type=content_type, **extra)
         if follow:
-            response = self._handle_redirects(response, **extra)
+            response = self._handle_redirects(response, data=data, format=format, content_type=content_type, **extra)
         return response
 
     def options(self, path, data=None, format=None, content_type=None,
@@ -328,7 +331,7 @@ class APIClient(APIRequestFactory, DjangoClient):
         response = super().options(
             path, data=data, format=format, content_type=content_type, **extra)
         if follow:
-            response = self._handle_redirects(response, **extra)
+            response = self._handle_redirects(response, data=data, format=format, content_type=content_type, **extra)
         return response
 
     def logout(self):
@@ -394,19 +397,7 @@ class URLPatternsTestCase(testcases.SimpleTestCase):
 
         cls._override.enable()
 
-        if django.VERSION > (4, 0):
-            cls.addClassCleanup(cls._override.disable)
-            cls.addClassCleanup(cleanup_url_patterns, cls)
+        cls.addClassCleanup(cls._override.disable)
+        cls.addClassCleanup(cleanup_url_patterns, cls)
 
         super().setUpClass()
-
-    if django.VERSION < (4, 0):
-        @classmethod
-        def tearDownClass(cls):
-            super().tearDownClass()
-            cls._override.disable()
-
-            if hasattr(cls, '_module_urlpatterns'):
-                cls._module.urlpatterns = cls._module_urlpatterns
-            else:
-                del cls._module.urlpatterns
