@@ -8,7 +8,7 @@ from django.test import TestCase
 from django.test.utils import override_settings
 from django.urls import path
 
-from rest_framework.compat import uritemplate, yaml
+from rest_framework.compat import coreapi, uritemplate, yaml
 from rest_framework.management.commands import generateschema
 from rest_framework.utils import formatting, json
 from rest_framework.views import APIView
@@ -52,9 +52,9 @@ class GenerateSchemaTests(TestCase):
     @pytest.mark.skipif(yaml is None, reason='PyYAML is required.')
     def test_renders_default_schema_with_custom_title_url_and_description(self):
         call_command('generateschema',
-                     '--title=SampleAPI',
-                     '--url=http://api.sample.com',
-                     '--description=Sample description',
+                     '--title=ExampleAPI',
+                     '--url=http://api.example.com',
+                     '--description=Example description',
                      stdout=self.out)
         # Check valid YAML was output.
         schema = yaml.safe_load(self.out.getvalue())
@@ -91,11 +91,12 @@ class GenerateSchemaTests(TestCase):
             os.remove(path)
 
     @pytest.mark.skipif(yaml is None, reason='PyYAML is required.')
+    @pytest.mark.skipif(coreapi is None, reason='coreapi is required.')
     @override_settings(REST_FRAMEWORK={'DEFAULT_SCHEMA_CLASS': 'rest_framework.schemas.AutoSchema'})
     def test_coreapi_renders_default_schema_with_custom_title_url_and_description(self):
         expected_out = """info:
-                            description: Sample description
-                            title: SampleAPI
+                            description: Example description
+                            title: ExampleAPI
                             version: ''
                           openapi: 3.0.0
                           paths:
@@ -103,16 +104,17 @@ class GenerateSchemaTests(TestCase):
                               get:
                                 operationId: list
                           servers:
-                          - url: http://api.sample.com/
+                          - url: http://api.example.com/
                           """
         call_command('generateschema',
-                     '--title=SampleAPI',
-                     '--url=http://api.sample.com',
-                     '--description=Sample description',
+                     '--title=ExampleAPI',
+                     '--url=http://api.example.com',
+                     '--description=Example description',
                      stdout=self.out)
 
         self.assertIn(formatting.dedent(expected_out), self.out.getvalue())
 
+    @pytest.mark.skipif(coreapi is None, reason='coreapi is required.')
     @override_settings(REST_FRAMEWORK={'DEFAULT_SCHEMA_CLASS': 'rest_framework.schemas.AutoSchema'})
     def test_coreapi_renders_openapi_json_schema(self):
         expected_out = {
@@ -142,6 +144,7 @@ class GenerateSchemaTests(TestCase):
 
         self.assertDictEqual(out_json, expected_out)
 
+    @pytest.mark.skipif(coreapi is None, reason='coreapi is required.')
     @override_settings(REST_FRAMEWORK={'DEFAULT_SCHEMA_CLASS': 'rest_framework.schemas.AutoSchema'})
     def test_renders_corejson_schema(self):
         expected_out = """{"_type":"document","":{"list":{"_type":"link","url":"/","action":"get"}}}"""
