@@ -5,6 +5,7 @@ import pytest
 from django.db import models
 from django.test import RequestFactory, TestCase, override_settings
 from django.urls import path
+from django.utils.safestring import SafeString
 from django.utils.translation import gettext_lazy as _
 
 from rest_framework import filters, generics, pagination, routers, serializers
@@ -18,6 +19,7 @@ from rest_framework.renderers import (
 from rest_framework.request import Request
 from rest_framework.schemas.openapi import AutoSchema, SchemaGenerator
 
+from ..models import BasicModel
 from . import views
 
 
@@ -111,6 +113,20 @@ class TestFieldMapping(TestCase):
         assert data['properties']['default_false']['default'] is False, "default must be false"
         assert 'default' not in data['properties']['without_default'], "default must not be defined"
 
+    def test_custom_field_name(self):
+        class CustomSchema(AutoSchema):
+            def get_field_name(self, field):
+                return 'custom_' + field.field_name
+
+        class Serializer(serializers.Serializer):
+            text_field = serializers.CharField()
+
+        inspector = CustomSchema()
+
+        data = inspector.map_serializer(Serializer())
+        assert 'custom_text_field' in data['properties']
+        assert 'text_field' not in data['properties']
+
     def test_nullable_fields(self):
         class Model(models.Model):
             rw_field = models.CharField(null=True)
@@ -128,6 +144,22 @@ class TestFieldMapping(TestCase):
         assert data['properties']['rw_field']['nullable'], "rw_field nullable must be true"
         assert data['properties']['ro_field']['nullable'], "ro_field nullable must be true"
         assert data['properties']['ro_field']['readOnly'], "ro_field read_only must be true"
+
+    def test_primary_key_related_field(self):
+        class PrimaryKeyRelatedFieldSerializer(serializers.Serializer):
+            basic = serializers.PrimaryKeyRelatedField(queryset=BasicModel.objects.all())
+            uuid = serializers.PrimaryKeyRelatedField(queryset=BasicModel.objects.all(),
+                                                      pk_field=serializers.UUIDField())
+            char = serializers.PrimaryKeyRelatedField(queryset=BasicModel.objects.all(),
+                                                      pk_field=serializers.CharField())
+
+        serializer = PrimaryKeyRelatedFieldSerializer()
+        inspector = AutoSchema()
+
+        data = inspector.map_serializer(serializer=serializer)
+        assert data['properties']['basic']['type'] == "integer"
+        assert data['properties']['uuid']['format'] == "uuid"
+        assert data['properties']['char']['type'] == "string"
 
 
 @pytest.mark.skipif(uritemplate is None, reason='uritemplate not installed.')
@@ -371,6 +403,56 @@ class TestOperationIntrospection(TestCase):
         assert list(schema['properties']['nested']['properties'].keys()) == ['number']
         assert schema['properties']['nested']['required'] == ['number']
 
+    def test_response_body_partial_serializer(self):
+        path = '/'
+        method = 'GET'
+
+        class ItemSerializer(serializers.Serializer):
+            text = serializers.CharField()
+
+            def __init__(self, *args, **kwargs):
+                super().__init__(*args, **kwargs)
+                self.partial = True
+
+        class View(generics.GenericAPIView):
+            serializer_class = ItemSerializer
+
+        view = create_view(
+            View,
+            method,
+            create_request(path),
+        )
+        inspector = AutoSchema()
+        inspector.view = view
+
+        responses = inspector.get_responses(path, method)
+        assert responses == {
+            '200': {
+                'description': '',
+                'content': {
+                    'application/json': {
+                        'schema': {
+                            'type': 'array',
+                            'items': {
+                                '$ref': '#/components/schemas/Item'
+                            },
+                        },
+                    },
+                },
+            },
+        }
+        components = inspector.get_components(path, method)
+        assert components == {
+            'Item': {
+                'type': 'object',
+                'properties': {
+                    'text': {
+                        'type': 'string',
+                    },
+                },
+            }
+        }
+
     def test_list_response_body_generation(self):
         """Test that an array schema is returned for list views."""
         path = '/'
@@ -569,6 +651,11 @@ class TestOperationIntrospection(TestCase):
             renderer.render(data) == b'o2:\n  test: test\no1:\n  test: test\n'  # py <= 3.5
         )
 
+    def test_openapi_yaml_safestring_render(self):
+        renderer = OpenAPIRenderer()
+        data = {'o1': SafeString('test')}
+        assert renderer.render(data) == b'o1: test\n'
+
     def test_serializer_filefield(self):
         path = '/{id}/'
         method = 'POST'
@@ -694,6 +781,21 @@ class TestOperationIntrospection(TestCase):
 
         operationId = inspector.get_operation_id(path, method)
         assert operationId == 'listUlysses'
+
+    def test_operation_id_plural(self):
+        path = '/'
+        method = 'GET'
+
+        view = create_view(
+            views.ExampleGenericAPIView,
+            method,
+            create_request(path),
+        )
+        inspector = AutoSchema(operation_id_base='City')
+        inspector.view = view
+
+        operationId = inspector.get_operation_id(path, method)
+        assert operationId == 'listCities'
 
     def test_operation_id_override_get(self):
         class CustomSchema(AutoSchema):
@@ -1110,6 +1212,31 @@ class TestGenerator(TestCase):
         assert b'"openapi": "' in ret
         assert b'"default": "0.0"' in ret
 
+    def test_schema_rendering_to_yaml(self):
+        patterns = [
+            path('example/', views.ExampleGenericAPIView.as_view()),
+        ]
+        generator = SchemaGenerator(patterns=patterns)
+
+        request = create_request('/')
+        schema = generator.get_schema(request=request)
+        ret = OpenAPIRenderer().render(schema)
+        assert b"openapi: " in ret
+        assert b"default: '0.0'" in ret
+
+    def test_schema_rendering_timedelta_to_yaml_with_validator(self):
+
+        patterns = [
+            path('example/', views.ExampleValidatedAPIView.as_view()),
+        ]
+        generator = SchemaGenerator(patterns=patterns)
+
+        request = create_request('/')
+        schema = generator.get_schema(request=request)
+        ret = OpenAPIRenderer().render(schema)
+        assert b"openapi: " in ret
+        assert b"duration:\n          type: string\n          minimum: \'10.0\'\n" in ret
+
     def test_schema_with_no_paths(self):
         patterns = []
         generator = SchemaGenerator(patterns=patterns)
@@ -1220,7 +1347,7 @@ class TestGenerator(TestCase):
 
             assert len(w) == 1
             assert issubclass(w[-1].category, UserWarning)
-            assert 'has been overriden with a different value.' in str(w[-1].message)
+            assert 'has been overridden with a different value.' in str(w[-1].message)
 
         assert 'components' in schema
         assert 'schemas' in schema['components']
