@@ -8,9 +8,9 @@ an appropriate set of serializer fields for each case.
 import datetime
 import decimal
 import json  # noqa
+import re
 import sys
 import tempfile
-from collections import OrderedDict
 
 import pytest
 from django.core.exceptions import ImproperlyConfigured
@@ -169,33 +169,32 @@ class TestRegularFieldMappings(TestCase):
                 model = RegularFieldsModel
                 fields = '__all__'
 
-        expected = dedent("""
-            TestSerializer():
-                auto_field = IntegerField(read_only=True)
-                big_integer_field = IntegerField()
-                boolean_field = BooleanField(required=False)
-                char_field = CharField(max_length=100)
-                comma_separated_integer_field = CharField(max_length=100, validators=[<django.core.validators.RegexValidator object>])
-                date_field = DateField()
-                datetime_field = DateTimeField()
-                decimal_field = DecimalField(decimal_places=1, max_digits=3)
-                email_field = EmailField(max_length=100)
-                float_field = FloatField()
-                integer_field = IntegerField()
-                null_boolean_field = BooleanField(allow_null=True, required=False)
-                positive_integer_field = IntegerField()
-                positive_small_integer_field = IntegerField()
-                slug_field = SlugField(allow_unicode=False, max_length=100)
-                small_integer_field = IntegerField()
-                text_field = CharField(max_length=100, style={'base_template': 'textarea.html'})
-                file_field = FileField(max_length=100)
-                time_field = TimeField()
-                url_field = URLField(max_length=100)
-                custom_field = ModelField(model_field=<tests.test_model_serializer.CustomField: custom_field>)
-                file_path_field = FilePathField(path=%r)
+        expected = dedent(r"""
+            TestSerializer\(\):
+                auto_field = IntegerField\(read_only=True\)
+                big_integer_field = IntegerField\(.*\)
+                boolean_field = BooleanField\(required=False\)
+                char_field = CharField\(max_length=100\)
+                comma_separated_integer_field = CharField\(max_length=100, validators=\[<django.core.validators.RegexValidator object>\]\)
+                date_field = DateField\(\)
+                datetime_field = DateTimeField\(\)
+                decimal_field = DecimalField\(decimal_places=1, max_digits=3\)
+                email_field = EmailField\(max_length=100\)
+                float_field = FloatField\(\)
+                integer_field = IntegerField\(.*\)
+                null_boolean_field = BooleanField\(allow_null=True, required=False\)
+                positive_integer_field = IntegerField\(.*\)
+                positive_small_integer_field = IntegerField\(.*\)
+                slug_field = SlugField\(allow_unicode=False, max_length=100\)
+                small_integer_field = IntegerField\(.*\)
+                text_field = CharField\(max_length=100, style={'base_template': 'textarea.html'}\)
+                file_field = FileField\(max_length=100\)
+                time_field = TimeField\(\)
+                url_field = URLField\(max_length=100\)
+                custom_field = ModelField\(model_field=<tests.test_model_serializer.CustomField: custom_field>\)
+                file_path_field = FilePathField\(path=%r\)
         """ % tempfile.gettempdir())
-
-        self.assertEqual(repr(TestSerializer()), expected)
+        assert re.search(expected, repr(TestSerializer())) is not None
 
     def test_field_options(self):
         class TestSerializer(serializers.ModelSerializer):
@@ -203,19 +202,19 @@ class TestRegularFieldMappings(TestCase):
                 model = FieldOptionsModel
                 fields = '__all__'
 
-        expected = dedent("""
-            TestSerializer():
-                id = IntegerField(label='ID', read_only=True)
-                value_limit_field = IntegerField(max_value=10, min_value=1)
-                length_limit_field = CharField(max_length=12, min_length=3)
-                blank_field = CharField(allow_blank=True, max_length=10, required=False)
-                null_field = IntegerField(allow_null=True, required=False)
-                default_field = IntegerField(required=False)
-                descriptive_field = IntegerField(help_text='Some help text', label='A label')
-                choices_field = ChoiceField(choices=(('red', 'Red'), ('blue', 'Blue'), ('green', 'Green')))
-                text_choices_field = ChoiceField(choices=(('red', 'Red'), ('blue', 'Blue'), ('green', 'Green')))
+        expected = dedent(r"""
+            TestSerializer\(\):
+                id = IntegerField\(label='ID', read_only=True\)
+                value_limit_field = IntegerField\(max_value=10, min_value=1\)
+                length_limit_field = CharField\(max_length=12, min_length=3\)
+                blank_field = CharField\(allow_blank=True, max_length=10, required=False\)
+                null_field = IntegerField\(allow_null=True,.*required=False\)
+                default_field = IntegerField\(.*required=False\)
+                descriptive_field = IntegerField\(help_text='Some help text', label='A label'.*\)
+                choices_field = ChoiceField\(choices=(?:\[|\()\('red', 'Red'\), \('blue', 'Blue'\), \('green', 'Green'\)(?:\]|\))\)
+                text_choices_field = ChoiceField\(choices=(?:\[|\()\('red', 'Red'\), \('blue', 'Blue'\), \('green', 'Green'\)(?:\]|\))\)
         """)
-        self.assertEqual(repr(TestSerializer()), expected)
+        assert re.search(expected, repr(TestSerializer())) is not None
 
     def test_nullable_boolean_field_choices(self):
         class NullableBooleanChoicesModel(models.Model):
@@ -315,7 +314,8 @@ class TestRegularFieldMappings(TestCase):
                 model = RegularFieldsModel
                 fields = ('auto_field', 'invalid')
 
-        expected = 'Field name `invalid` is not valid for model `RegularFieldsModel`.'
+        expected = 'Field name `invalid` is not valid for model `RegularFieldsModel` ' \
+                   'in `tests.test_model_serializer.TestSerializer`.'
         with self.assertRaisesMessage(ImproperlyConfigured, expected):
             TestSerializer().fields
 
@@ -454,8 +454,8 @@ class TestPosgresFieldsMapping(TestCase):
 
         expected = dedent("""
             TestSerializer():
-                array_field = ListField(allow_empty=False, child=CharField(label='Array field', validators=[<django.core.validators.MaxLengthValidator object>]))
-                array_field_with_blank = ListField(child=CharField(label='Array field with blank', validators=[<django.core.validators.MaxLengthValidator object>]), required=False)
+                array_field = ListField(allow_empty=False, child=CharField(label='Array field'))
+                array_field_with_blank = ListField(child=CharField(label='Array field with blank'), required=False)
         """)
         self.assertEqual(repr(TestSerializer()), expected)
 
@@ -758,7 +758,7 @@ class TestRelationalFieldDisplayValue(TestCase):
                 fields = '__all__'
 
         serializer = TestSerializer()
-        expected = OrderedDict([(1, 'Red Color'), (2, 'Yellow Color'), (3, 'Green Color')])
+        expected = {1: 'Red Color', 2: 'Yellow Color', 3: 'Green Color'}
         self.assertEqual(serializer.fields['color'].choices, expected)
 
     def test_custom_display_value(self):
@@ -774,7 +774,7 @@ class TestRelationalFieldDisplayValue(TestCase):
                 fields = '__all__'
 
         serializer = TestSerializer()
-        expected = OrderedDict([(1, 'My Red Color'), (2, 'My Yellow Color'), (3, 'My Green Color')])
+        expected = {1: 'My Red Color', 2: 'My Yellow Color', 3: 'My Green Color'}
         self.assertEqual(serializer.fields['color'].choices, expected)
 
 
@@ -1021,6 +1021,73 @@ class Issue2704TestCase(TestCase):
         assert serializer.data == expected
 
 
+class Issue7550FooModel(models.Model):
+    text = models.CharField(max_length=100)
+    bar = models.ForeignKey(
+        'Issue7550BarModel', null=True, blank=True, on_delete=models.SET_NULL,
+        related_name='foos', related_query_name='foo')
+
+
+class Issue7550BarModel(models.Model):
+    pass
+
+
+class Issue7550TestCase(TestCase):
+
+    def test_dotted_source(self):
+
+        class _FooSerializer(serializers.ModelSerializer):
+            class Meta:
+                model = Issue7550FooModel
+                fields = ('id', 'text')
+
+        class FooSerializer(serializers.ModelSerializer):
+            other_foos = _FooSerializer(source='bar.foos', many=True)
+
+            class Meta:
+                model = Issue7550BarModel
+                fields = ('id', 'other_foos')
+
+        bar = Issue7550BarModel.objects.create()
+        foo_a = Issue7550FooModel.objects.create(bar=bar, text='abc')
+        foo_b = Issue7550FooModel.objects.create(bar=bar, text='123')
+
+        assert FooSerializer(foo_a).data == {
+            'id': foo_a.id,
+            'other_foos': [
+                {
+                    'id': foo_a.id,
+                    'text': foo_a.text,
+                },
+                {
+                    'id': foo_b.id,
+                    'text': foo_b.text,
+                },
+            ],
+        }
+
+    def test_dotted_source_with_default(self):
+
+        class _FooSerializer(serializers.ModelSerializer):
+            class Meta:
+                model = Issue7550FooModel
+                fields = ('id', 'text')
+
+        class FooSerializer(serializers.ModelSerializer):
+            other_foos = _FooSerializer(source='bar.foos', default=[], many=True)
+
+            class Meta:
+                model = Issue7550FooModel
+                fields = ('id', 'other_foos')
+
+        foo = Issue7550FooModel.objects.create(bar=None, text='abc')
+
+        assert FooSerializer(foo).data == {
+            'id': foo.id,
+            'other_foos': [],
+        }
+
+
 class DecimalFieldModel(models.Model):
     decimal_field = models.DecimalField(
         max_digits=3,
@@ -1263,12 +1330,12 @@ class TestFieldSource(TestCase):
                     }
                 }
 
-        expected = dedent("""
-            TestSerializer():
-                number_field = IntegerField(source='integer_field')
+        expected = dedent(r"""
+            TestSerializer\(\):
+                number_field = IntegerField\(.*source='integer_field'\)
         """)
         self.maxDiff = None
-        self.assertEqual(repr(TestSerializer()), expected)
+        assert re.search(expected, repr(TestSerializer())) is not None
 
 
 class Issue6110TestModel(models.Model):
