@@ -10,13 +10,20 @@ from rest_framework import permissions, serializers, viewsets
 from rest_framework.decorators import action
 from rest_framework.response import Response
 from rest_framework.routers import DefaultRouter, SimpleRouter
-from rest_framework.test import APIRequestFactory, URLPatternsTestCase
+from rest_framework.test import (
+    APIClient, APIRequestFactory, URLPatternsTestCase
+)
 from rest_framework.utils import json
 
 factory = APIRequestFactory()
 
 
 class RouterTestModel(models.Model):
+    uuid = models.CharField(max_length=20)
+    text = models.CharField(max_length=200)
+
+
+class BasenameTestModel(models.Model):
     uuid = models.CharField(max_length=20)
     text = models.CharField(max_length=200)
 
@@ -40,6 +47,11 @@ class KWargedNoteViewSet(viewsets.ModelViewSet):
     serializer_class = NoteSerializer
     lookup_field = 'text__contains'
     lookup_url_kwarg = 'text'
+
+
+class BasenameViewSet(viewsets.ModelViewSet):
+    queryset = BasenameTestModel.objects.all()
+    serializer_class = None
 
 
 class MockViewSet(viewsets.ModelViewSet):
@@ -75,8 +87,34 @@ class RegexUrlPathViewSet(viewsets.ViewSet):
         return Response({'pk': pk, 'kwarg': kwarg})
 
 
+class UrlPathViewSet(viewsets.ViewSet):
+    @action(detail=False, url_path='list/<int:kwarg>')
+    def url_path_list(self, request, *args, **kwargs):
+        kwarg = self.kwargs.get('kwarg', '')
+        return Response({'kwarg': kwarg})
+
+    @action(detail=True, url_path='detail/<int:kwarg>')
+    def url_path_detail(self, request, *args, **kwargs):
+        pk = self.kwargs.get('pk', '')
+        kwarg = self.kwargs.get('kwarg', '')
+        return Response({'pk': pk, 'kwarg': kwarg})
+
+    @action(detail=True, url_path='detail/<int:kwarg>/detail/<int:param>')
+    def url_path_detail_multiple_params(self, request, *args, **kwargs):
+        pk = self.kwargs.get('pk', '')
+        kwarg = self.kwargs.get('kwarg', '')
+        param = self.kwargs.get('param', '')
+        return Response({'pk': pk, 'kwarg': kwarg, 'param': param})
+
+
 notes_router = SimpleRouter()
 notes_router.register(r'notes', NoteViewSet)
+
+notes_path_router = SimpleRouter(use_regex_path=False)
+notes_path_router.register('notes', NoteViewSet)
+
+notes_path_default_router = DefaultRouter(use_regex_path=False)
+notes_path_default_router.register('notes', NoteViewSet)
 
 kwarged_notes_router = SimpleRouter()
 kwarged_notes_router.register(r'notes', KWargedNoteViewSet)
@@ -89,6 +127,9 @@ empty_prefix_router.register(r'', EmptyPrefixViewSet, basename='empty_prefix')
 
 regex_url_path_router = SimpleRouter()
 regex_url_path_router.register(r'', RegexUrlPathViewSet, basename='regex')
+
+url_path_router = SimpleRouter(use_regex_path=False)
+url_path_router.register('', UrlPathViewSet, basename='path')
 
 
 class BasicViewSet(viewsets.ViewSet):
@@ -156,7 +197,7 @@ class TestSimpleRouter(URLPatternsTestCase, TestCase):
     def test_register_after_accessing_urls(self):
         self.router.register(r'notes', NoteViewSet)
         assert len(self.router.urls) == 2  # list and detail
-        self.router.register(r'notes_bis', NoteViewSet)
+        self.router.register(r'notes_bis', NoteViewSet, basename='notes_bis')
         assert len(self.router.urls) == 4
 
 
@@ -459,6 +500,93 @@ class TestRegexUrlPath(URLPatternsTestCase, TestCase):
         assert json.loads(response.content.decode()) == {'pk': pk, 'kwarg': kwarg}
 
 
+class TestUrlPath(URLPatternsTestCase, TestCase):
+    client_class = APIClient
+    urlpatterns = [
+        path('path/', include(url_path_router.urls)),
+        path('default/', include(notes_path_default_router.urls)),
+        path('example/', include(notes_path_router.urls)),
+    ]
+
+    def setUp(self):
+        RouterTestModel.objects.create(uuid='123', text='foo bar')
+        RouterTestModel.objects.create(uuid='a b', text='baz qux')
+
+    def test_create(self):
+        new_note = {
+            'uuid': 'foo',
+            'text': 'example'
+        }
+        response = self.client.post('/example/notes/', data=new_note)
+        assert response.status_code == 201
+        assert response['location'] == 'http://testserver/example/notes/foo/'
+        assert response.data == {"url": "http://testserver/example/notes/foo/", "uuid": "foo", "text": "example"}
+        assert RouterTestModel.objects.filter(uuid='foo').exists()
+
+    def test_retrieve(self):
+        for url in ('/example/notes/123/', '/default/notes/123/'):
+            with self.subTest(url=url):
+                response = self.client.get(url)
+                assert response.status_code == 200
+                # only gets example path since was the last to be registered
+                assert response.data == {"url": "http://testserver/example/notes/123/", "uuid": "123", "text": "foo bar"}
+
+    def test_list(self):
+        for url in ('/example/notes/', '/default/notes/'):
+            with self.subTest(url=url):
+                response = self.client.get(url)
+                assert response.status_code == 200
+                # only gets example path since was the last to be registered
+                assert response.data == [
+                    {"url": "http://testserver/example/notes/123/", "uuid": "123", "text": "foo bar"},
+                    {"url": "http://testserver/example/notes/a%20b/", "uuid": "a b", "text": "baz qux"},
+                ]
+
+    def test_update(self):
+        updated_note = {
+            'text': 'foo bar example'
+        }
+        response = self.client.patch('/example/notes/123/', data=updated_note)
+        assert response.status_code == 200
+        assert response.data == {"url": "http://testserver/example/notes/123/", "uuid": "123", "text": "foo bar example"}
+
+    def test_delete(self):
+        response = self.client.delete('/example/notes/123/')
+        assert response.status_code == 204
+        assert not RouterTestModel.objects.filter(uuid='123').exists()
+
+    def test_list_extra_action(self):
+        kwarg = 1234
+        response = self.client.get('/path/list/{}/'.format(kwarg))
+        assert response.status_code == 200
+        assert json.loads(response.content.decode()) == {'kwarg': kwarg}
+
+    def test_detail_extra_action(self):
+        pk = '1'
+        kwarg = 1234
+        response = self.client.get('/path/{}/detail/{}/'.format(pk, kwarg))
+        assert response.status_code == 200
+        assert json.loads(response.content.decode()) == {'pk': pk, 'kwarg': kwarg}
+
+    def test_detail_extra_other_action(self):
+        # this to assure that ambiguous patterns are interpreted correctly
+        # using the `path` converters this URL is recognized to match the pattern
+        # of `UrlPathViewSet.url_path_detail` when it should match
+        # `UrlPathViewSet.url_path_detail_multiple_params`
+        pk = '1'
+        kwarg = 1234
+        param = 2
+        response = self.client.get('/path/1/detail/1234/detail/2/')
+        assert response.status_code == 200
+        assert json.loads(response.content.decode()) == {'pk': pk, 'kwarg': kwarg, 'param': param}
+
+    def test_defaultrouter_root(self):
+        response = self.client.get('/default/')
+        assert response.status_code == 200
+        # only gets example path since was the last to be registered
+        assert response.data == {"notes": "http://testserver/example/notes/"}
+
+
 class TestViewInitkwargs(URLPatternsTestCase, TestCase):
     urlpatterns = [
         path('example/', include(notes_router.urls)),
@@ -481,3 +609,103 @@ class TestViewInitkwargs(URLPatternsTestCase, TestCase):
         initkwargs = match.func.initkwargs
 
         assert initkwargs['basename'] == 'routertestmodel'
+
+
+class BasenameTestCase:
+    def test_conflicting_autogenerated_basenames(self):
+        """
+        Ensure 2 routers with the same model, and no basename specified
+        throws an ImproperlyConfigured exception
+        """
+        self.router.register(r'notes', NoteViewSet)
+
+        with pytest.raises(ImproperlyConfigured):
+            self.router.register(r'notes_kwduplicate', KWargedNoteViewSet)
+
+        with pytest.raises(ImproperlyConfigured):
+            self.router.register(r'notes_duplicate', NoteViewSet)
+
+    def test_conflicting_mixed_basenames(self):
+        """
+        Ensure 2 routers with the same model, and no basename specified on 1
+        throws an ImproperlyConfigured exception
+        """
+        self.router.register(r'notes', NoteViewSet)
+
+        with pytest.raises(ImproperlyConfigured):
+            self.router.register(r'notes_kwduplicate', KWargedNoteViewSet, basename='routertestmodel')
+
+        with pytest.raises(ImproperlyConfigured):
+            self.router.register(r'notes_duplicate', NoteViewSet, basename='routertestmodel')
+
+    def test_nonconflicting_mixed_basenames(self):
+        """
+        Ensure 2 routers with the same model, and a distinct basename
+        specified on the second router does not fail
+        """
+        self.router.register(r'notes', NoteViewSet)
+        self.router.register(r'notes_kwduplicate', KWargedNoteViewSet, basename='routertestmodel_kwduplicate')
+        self.router.register(r'notes_duplicate', NoteViewSet, basename='routertestmodel_duplicate')
+
+    def test_conflicting_specified_basename(self):
+        """
+        Ensure 2 routers with the same model, and the same basename specified
+        on both throws an ImproperlyConfigured exception
+        """
+        self.router.register(r'notes', NoteViewSet, basename='notes')
+
+        with pytest.raises(ImproperlyConfigured):
+            self.router.register(r'notes_kwduplicate', KWargedNoteViewSet, basename='notes')
+
+        with pytest.raises(ImproperlyConfigured):
+            self.router.register(r'notes_duplicate', KWargedNoteViewSet, basename='notes')
+
+    def test_nonconflicting_specified_basename(self):
+        """
+        Ensure 2 routers with the same model, and a distinct basename specified
+        on each does not throw an exception
+        """
+        self.router.register(r'notes', NoteViewSet, basename='notes')
+        self.router.register(r'notes_kwduplicate', KWargedNoteViewSet, basename='notes_kwduplicate')
+        self.router.register(r'notes_duplicate', NoteViewSet, basename='notes_duplicate')
+
+    def test_nonconflicting_specified_basename_different_models(self):
+        """
+        Ensure 2 routers with different models, and a distinct basename specified
+        on each does not throw an exception
+        """
+        self.router.register(r'notes', NoteViewSet, basename='notes')
+        self.router.register(r'notes_basename', BasenameViewSet, basename='notes_basename')
+
+    def test_conflicting_specified_basename_different_models(self):
+        """
+        Ensure 2 routers with different models, and a conflicting basename specified
+        throws an exception
+        """
+        self.router.register(r'notes', NoteViewSet)
+        with pytest.raises(ImproperlyConfigured):
+            self.router.register(r'notes_basename', BasenameViewSet, basename='routertestmodel')
+
+    def test_nonconflicting_autogenerated_basename_different_models(self):
+        """
+        Ensure 2 routers with different models, and a distinct basename specified
+        on each does not throw an exception
+        """
+        self.router.register(r'notes', NoteViewSet)
+        self.router.register(r'notes_basename', BasenameViewSet)
+
+
+class TestDuplicateBasenameSimpleRouter(BasenameTestCase, TestCase):
+    def setUp(self):
+        self.router = SimpleRouter(trailing_slash=False)
+
+
+class TestDuplicateBasenameDefaultRouter(BasenameTestCase, TestCase):
+    def setUp(self):
+        self.router = DefaultRouter()
+
+
+class TestDuplicateBasenameDefaultRouterRootViewName(BasenameTestCase, TestCase):
+    def setUp(self):
+        self.router = DefaultRouter()
+        self.router.root_view_name = 'nameable-root'
